requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL + PostGIS
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
alembic==1.13.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# HTTP Client for geocoding APIs
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
shapely==2.1.2
